refactored /helpers into /helpers and /controllers
