+++ helpers/libraries/errorHandlers.js
@@ -1,5 +1,3 @@
-var path = require('path');
-
 module.exports = {
 	handleRequestError: function(error, res) {
 		if ((error === "NO_CLAIMS") || (error === "NO_FREE_PUBLIC_CLAIMS")){
@@ -11,5 +9,14 @@ module.exports = {
 		} else {
 			res.status(400).send(error.toString());
 		};
-	}
+	},
+	handlePublishError: function(error) {
+	if (error.code === "ECONNREFUSED"){
+		return "Connection refused.  The daemon may not be running.";
+	} else if (error.response.data.error) {
+		return error.response.data.error;
+	} else {
+		return error;
+	};
+}
 }
