+++ helpers/libraries/lbryApi.js
@@ -1,5 +1,5 @@
 var axios = require('axios');
-var db = require("../models");
+var db = require("../../models");
 
 module.exports = {
 	publishClaim: function(publishParams, fileType){
@@ -9,19 +9,20 @@ module.exports = {
 				"method": "publish", 
 				"params": publishParams
 			}).then(function (response) {
-				console.log(">> 'publish' success", response);
+				console.log(">> 'publish' success");
+				var result = response.data.result;
 				db.File.create({
 					name: publishParams.name,
-					claim_id: response.data.result.claim_id,
-					outpoint: response.data.result.outpoint,
+					claim_id: result.claim_id,
+					outpoint: result.txid + ":" + result.nout,
 					file_name: "test",
-					file_path: publishParams.filePath,
+					file_path: publishParams.file_path,
 					file_type: fileType,
-					nsfw: nsfw,
+					nsfw: publishParams.metadata.nsfw,
 				}).catch(function(error){
 					console.log('An error occurred when writing to the MySQL database:', error);
 				});
-				resolve(response.data.result);
+				resolve(result);
 			}).catch(function(error){
 				console.log(">> 'publish' error");
 				reject(error);
@@ -36,7 +37,7 @@ module.exports = {
 				"method": "get",
 				"params": { "uri": uri, "timeout": 20}
 			}).then(function (response) {
-				console.log(">> 'get' success", response.data);
+				console.log(">> 'get' success");
 				//check to make sure the daemon didn't just time out
 				if (response.data.result.error){
 					reject(response.data.result.error);
@@ -45,24 +46,22 @@ module.exports = {
 					note: put in a check to make sure we do not resolve until the download is actually complete (response.data.completed === true)
 				*/
 				// save a record of the file to the Files table
+				var result = response.data.result
 				db.File.create({
-					name: response.data.result.name,
-					claim_id: response.data.result.claim_id,
-					outpoint: response.data.result.outpoint,
-					file_name: response.data.result.file_name,
-					file_path: response.data.result.download_path,
-					file_type: response.data.result.mime_type,
-					nsfw: response.data.result.metadata.stream.metadata.nsfw,
+					name: result.name,
+					claim_id: result.claim_id,
+					outpoint: result.outpoint,
+					file_name: result.file_name,
+					file_path: result.download_path,
+					file_type: result.mime_type,
+					nsfw: result.metadata.stream.metadata.nsfw,
 				}).catch(function(error){
 					console.log('An error occurred when writing to the MySQL database:', error);
 				});
-				// resolve the promise
-				resolve(response.data.result);
+				resolve(result);
 			}).catch(function(error){
 				console.log(">> 'get' error");
-				// reject the promise with an error message
 				reject(error);
-				return;
 			});
 		});
 		return deferred;
@@ -75,7 +74,7 @@ module.exports = {
 				params: { name: claimName }
 			}).then(function (response) {
 				console.log(">> 'claim_list' success");
-				resolve(response.data); 
+				resolve(response.data.result); 
 			}).catch(function(error){
 				console.log(">> 'claim_list' error");
 				reject(error);
