+++ routes/serve-routes.js
@@ -1,3 +1,6 @@
+var errorHandlers = require("../helpers/libraries/errorHandlers.js");
+var serveController = require("../controllers/serveController.js");
+
 function serveFile(fileInfo, res){
 	// set default options
 	var options = {
@@ -25,7 +28,7 @@ function serveFile(fileInfo, res){
 	res.status(200).sendFile(fileInfo.file_path, options);
 }
 
-module.exports = function(app, routeHelpers, lbryHelpers, ua, googleAnalyticsId){
+module.exports = function(app, ua, googleAnalyticsId){
 	// route to fetch one free public claim 
 	app.get("/:name/:claim_id", function(req, res){
 		var routeString = req.params.name + "/" + req.params.claim_id;
@@ -33,13 +36,13 @@ module.exports = function(app, routeHelpers, lbryHelpers, ua, googleAnalyticsId)
 		ua(googleAnalyticsId, {https: true}).event("Serve Route", "/name/claimId", routeString).send();
 		// begin image-serve processes
 		console.log(">> GET request on /" + routeString);
-		lbryHelpers.getClaimBasedOnUri(req.params.name, req.params.claim_id)
+		serveController.getClaimByClaimId(req.params.name, req.params.claim_id)
 		.then(function(fileInfo){
 			console.log("/:name/:claim_id success.", fileInfo.file_name);
 			serveFile(fileInfo, res);
 		}).catch(function(error){
 			console.log("/:name/:claim_id error:", error)
-			routeHelpers.handleRequestError(error, res);
+			errorHandlers.handleRequestError(error, res);
 		});
 	});
 	// route to fetch one free public claim 
@@ -48,14 +51,13 @@ module.exports = function(app, routeHelpers, lbryHelpers, ua, googleAnalyticsId)
 		ua(googleAnalyticsId, {https: true}).event("Serve Route", "/name", req.params.name).send();
 		// begin image-serve processes
 		console.log(">> GET request on /" + req.params.name);
-		lbryHelpers.getClaimBasedOnNameOnly(req.params.name)
+		serveController.getClaimByName(req.params.name)
 		.then(function(fileInfo){
-			
 			console.log("/:name success.", fileInfo.file_name);
 			serveFile(fileInfo, res);
 		}).catch(function(error){
 			console.log("/:name error:", error);
-			routeHelpers.handleRequestError(error, res);
+			errorHandlers.handleRequestError(error, res);
 		});
 	});
 }
