+++ routes/show-routes.js
@@ -1,11 +1,14 @@
+var errorHandlers = require("../helpers/libraries/errorHandlers.js");
+var showController = require("../controllers/serveController.js");
 
-module.exports = function(app, routeHelpers, lbryHelpers, ua, googleAnalyticsId){
-	// route to fetch all free public claims 
+module.exports = function(app, ua, googleAnalyticsId){
+	// route to fetch all free public claims
 	app.get("/:name/all", function(req, res){
 		console.log(">> GET request on /" + req.params.name + "/all");
+		// google analytics
 		ua(googleAnalyticsId, {https: true}).event("Show Routes", "/name/all", req.params.name + "/all").send();
-		// create promise
-		lbryHelpers.getAllClaims(req.params.name)
+		// fetch all free public claims
+		showController.getAllClaims(req.params.name)
 		.then(function(orderedFreePublicClaims){
 			console.log("/:name/all success.");
 			res.status(200).render('allClaims', { claims: orderedFreePublicClaims });
@@ -13,7 +16,7 @@ module.exports = function(app, routeHelpers, lbryHelpers, ua, googleAnalyticsId)
 		})
 		.catch(function(error){
 			console.log("/:name/all error:", error);
-			routeHelpers.handleRequestError(error, res);
+			errorHandlers.handleRequestError(error, res);
 		})
 	});
 }
