+++ routes/sockets-routes.js
@@ -1,6 +1,6 @@
-var socketHelpers = require('../helpers/socketHelpers.js');
+var publishController = require('../controllers/publishController.js');
 
-module.exports = function(app, path, siofu, hostedContentPath, ua, googleAnalyticsId) {
+module.exports = function(app, siofu, hostedContentPath, ua, googleAnalyticsId) {
 	var http = require('http').Server(app);
 	var io = require('socket.io')(http);
 	
@@ -10,7 +10,7 @@ module.exports = function(app, path, siofu, hostedContentPath, ua, googleAnalyti
 		var visitor = ua(googleAnalyticsId, {https: true});
 		// attach upload listeners
 		var uploader = new siofu();
-		uploader.dir = hostedContentPath; //path.join(__dirname, '../../Uploads');
+		uploader.dir = hostedContentPath;
 		uploader.listen(socket);
 		uploader.on("start", function(event){
 			// server side test to make sure file is not a bad file type
@@ -26,7 +26,7 @@ module.exports = function(app, path, siofu, hostedContentPath, ua, googleAnalyti
 			console.log("saved ", event.file.name, "deets:", event.file);
 			if (event.file.success){
 				socket.emit("publish-status", "file upload successfully completed");
-				socketHelpers.publish(event.file.meta.name, event.file.pathName, event.file.meta.type, event.file.meta.license, event.file.meta.nsfw, socket, visitor)
+				publishController.publish(event.file.meta.name, event.file.pathName, event.file.meta.type, event.file.meta.license, event.file.meta.nsfw, socket, visitor)
 			} else {
 				socket.emit("publish-failure", "file uploaded, but with errors")
 			};
